Extract `UIImage` test helper in its own file
